new: security config adjusted

+++ src/main/java/liqour/ravine/repositories/RatingRepository.java
@@ -1,6 +1,5 @@
 package liqour.ravine.repositories;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -11,8 +10,8 @@ import org.springframework.stereotype.Repository;
 import liqour.ravine.entities.Rating;
 
 @Repository
-public interface RatingRepository extends JpaRepository<Rating, BigInteger> {
+public interface RatingRepository extends JpaRepository<Rating, Integer> {
 
     @Query("SELECT r FROM Rating r WHERE r.rating_id = :id")
-    Optional<Rating> findRatingById(@Param("id") BigInteger id);
+    Optional<Rating> findRatingById(@Param("id") Integer id);
 }
